🚚 refactor: move config/app to env/app
