+++ src/app/(backend)/middleware/auth/utils.ts
@@ -1,7 +1,7 @@
 import { type AuthObject } from '@clerk/backend';
 
-import { getAppConfig } from '@/config/app';
 import { enableClerk, enableNextAuth } from '@/const/auth';
+import { getAppConfig } from '@/envs/app';
 import { AgentRuntimeError } from '@/libs/model-runtime';
 import { ChatErrorType } from '@/types/fetch';
