+++ src/utils/basePath.ts
@@ -1,3 +1,3 @@
-import { appEnv } from '@/config/app';
+import { appEnv } from '@/envs/app';
 
 export const withBasePath = (path: string) => appEnv.NEXT_PUBLIC_BASE_PATH + path;
